just everything at once
I totally forgot to actually commit my work to the master branch as i was going so i've basically just done most of this project all in one swoop as far as github is concerned.
So basically i've got all the 'meets' requirments down and now i'm working on doing the 'exceeds' portions.